requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[argon2]
langchain
langchain-text-splitters
langchain-openai
langchain-chroma
chromadb
pymupdf
aiosqlite
pydantic-settings
